+++ R/read.R
@@ -107,7 +107,22 @@ read_stars = function(.x, ..., options = character(0), driver = character(0),
 		} else
 			enc2utf8char(maybe_normalizePath(.x, np = normalize_path))
 
-	if (length(curvilinear) == 2 && is.character(curvilinear)) {
+	if (missing(curvilinear)) { # see if we can get it from metadat-item "GEOLOCATION":
+		geolocation <- try(gdal_metadata(.x, "GEOLOCATION"), silent = TRUE)
+		if (!inherits(geolocation, "try-error")) { # the thing is curvilinear:
+			gxy = if (packageVersion("sf") >= "1.0-8") # read coordinate arrays
+					c(geolocation$X_DATASET, geolocation$Y_DATASET)
+				else 
+					c(gdal_metadata(.x, c("GEOLOCATION", "X_DATASET"), parse = FALSE),
+						gdal_metadata(.x, c("GEOLOCATION", "Y_DATASET"), parse = FALSE))
+			lon = adrop(read_stars(gxy[1], driver = driver, quiet = quiet, NA_value = NA_value,
+				RasterIO = RasterIO, proxy = FALSE, curvilinear = character(0), ...))
+			lat = adrop(read_stars(gxy[2], driver = driver, quiet = quiet, NA_value = NA_value,
+				RasterIO = RasterIO, proxy = FALSE, curvilinear = character(0), ...))
+			curvilinear = setNames(c(st_set_dimensions(lon, names = c("x", "y")),
+				st_set_dimensions(lat, names = c("x", "y"))), c("x", "y"))
+		}
+	} else if (length(curvilinear) == 2 && is.character(curvilinear)) { # read them
 		lon = adrop(read_stars(.x, sub = curvilinear[1], driver = driver, quiet = quiet, NA_value = NA_value,
 			RasterIO = RasterIO, proxy = FALSE, ..., sub_required = TRUE))
 		lat = adrop(read_stars(.x, sub = curvilinear[2], driver = driver, quiet = quiet, NA_value = NA_value,